update: adding incremental support for MCP + nit fix for tests
