+++ pkg/http3/server.go
@@ -2,14 +2,9 @@ package http3
 
 import (
 	"context"
-	"fmt"
-	"log"
 	"net/http"
-	"strings"
 
-	"github.com/ayushanand18/as-http3lib/internal/config"
 	"github.com/ayushanand18/as-http3lib/internal/constants"
-	"github.com/ayushanand18/as-http3lib/internal/tls"
 	"github.com/ayushanand18/as-http3lib/internal/utils"
 	"github.com/ayushanand18/as-http3lib/pkg/types"
 	"github.com/gorilla/mux"
@@ -33,16 +28,21 @@ func (h *rootHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 }
 
 type server struct {
+	// hTTP server assets
 	qchttp3.Server
 	mux            *mux.Router
 	routeMatchMap  map[string]map[constants.HttpMethodTypes]types.HandlerFunc
 	http1ServerTLS http.Server
 	http1Server    http.Server
+
+	// MCP server assets
+	mcpToolsHandlerMap map[string]types.HandlerFunc
 }
 
 type Server interface {
 	Initialize(context.Context) error
 	ListenAndServe(context.Context) error
+	ListenAndServeMcp(context.Context) error
 
 	// HTTP Methods
 	GET(string) Method
@@ -54,6 +54,11 @@ type Server interface {
 	OPTIONS(string) Method
 	CONNECT(string) Method
 	TRACE(string) Method
+
+	// MCP Methods
+	MCP_TOOL(string) Method
+	MCP_RESOURCE(string) Method
+	MCP_PROMPT(string) Method
 }
 
 func NewServer(ctx context.Context) Server {
@@ -78,90 +83,4 @@ func NewServer(ctx context.Context) Server {
 		mux:           mux.NewRouter(),
 		routeMatchMap: make(map[string]map[constants.HttpMethodTypes]types.HandlerFunc),
 	}
-}
-
-func (s *server) Initialize(ctx context.Context) error {
-	if config.GetBool(ctx, "service.tls.generate_if_missing", true) && checkIfTlsCertificateIsMissing(ctx) {
-		if err := tls.GenerateSelfSignedCert(ctx); err != nil {
-			return fmt.Errorf("failed to generate self-signed certificate: %v", err)
-		}
-	}
-
-	tlsConfig := tls.GenerateTLSConfig(ctx)
-	root := &rootHandler{mux: s.mux}
-
-	s.Handler = root
-	s.TLSConfig = tlsConfig
-	s.TLSConfig.NextProtos = []string{"h3"}
-
-	h1Handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// if on H/1 advertise H/3
-		w.Header().Set("Alt-Svc", fmt.Sprintf(`h3=":%s"; ma=2592000,h3-29=":%s"; ma=2592000`, s.Addr[strings.LastIndex(s.Addr, ":")+1:], s.Addr[strings.LastIndex(s.Addr, ":")+1:]))
-		root.ServeHTTP(w, r)
-	})
-	s.http1Server.Handler = h1Handler
-	s.http1ServerTLS.Handler = h1Handler
-	s.http1ServerTLS.TLSConfig = tlsConfig
-
-	return nil
-}
-
-func (s *server) ListenAndServe(ctx context.Context) error {
-	utils.PrintStartBanner()
-	errChan := make(chan error, 2)
-
-	go func() {
-		log.Println("Starting HTTP/3 server on", s.Server.Addr)
-		errChan <- s.Server.ListenAndServe()
-	}()
-
-	go func() {
-		log.Println("Starting HTTP/1.1 + Alt-Svc server on", s.http1Server.Addr)
-		// server over http
-		errChan <- s.http1Server.ListenAndServe()
-	}()
-
-	go func() {
-		log.Println("Starting HTTP/1.1 + Alt-Svc server (HTTPS) on", s.http1ServerTLS.Addr)
-		// server over https
-		errChan <- s.http1ServerTLS.ListenAndServeTLS("", "")
-	}()
-
-	return <-errChan
-}
-
-func (s *server) GET(url string) Method {
-	return NewMethod(constants.HttpMethodGet, url, s)
-}
-
-func (s *server) POST(url string) Method {
-	return NewMethod(constants.HttpMethodPost, url, s)
-
-}
-func (s *server) PUT(url string) Method {
-	return NewMethod(constants.HttpMethodPut, url, s)
-}
-
-func (s *server) PATCH(url string) Method {
-	return NewMethod(constants.HttpMethodPatch, url, s)
-}
-
-func (s *server) DELETE(url string) Method {
-	return NewMethod(constants.HttpMethodDelete, url, s)
-}
-
-func (s *server) HEAD(url string) Method {
-	return NewMethod(constants.HttpMethodHead, url, s)
-}
-
-func (s *server) OPTIONS(url string) Method {
-	return NewMethod(constants.HttpMethodOptions, url, s)
-}
-
-func (s *server) CONNECT(url string) Method {
-	return NewMethod(constants.HttpMethodConnect, url, s)
-}
-
-func (s *server) TRACE(url string) Method {
-	return NewMethod(constants.HttpMethodTrace, url, s)
 }
